requirements.txt
gitpython>=3.1.18
matplotlib>=3.3.4
numba>=0.53.1
numpy>=1.19.5
pandas>=1.1.5
scipy>=1.5.4
scikit-learn>=0.24.2
xarray>=0.16.2
netcdf4>=1.5.7
psutil>=5.9.0
tqdm>=4.63.0
pyyaml>=6.0
addict>=2.4.0
